LPS-161984 Move ObjectFieldFormBase to new ObjectField directory
