+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/EditObjectStateField.tsx
@@ -21,7 +21,7 @@ import {
 import {openToast} from 'frontend-js-web';
 import React, {useEffect, useState} from 'react';
 
-import {useObjectFieldForm} from './ObjectFieldFormBase';
+import {useObjectFieldForm} from './ObjectField/ObjectFieldFormBase';
 import StateDefinition from './StateManager/StateDefinition';
 
 export default function EditObjectStateField({objectField, readOnly}: IProps) {
