+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ObjectField/SearchableContainer.tsx
@@ -16,7 +16,7 @@ import ClayForm, {ClayRadio, ClayRadioGroup, ClayToggle} from '@clayui/form';
 import {Card, Select} from '@liferay/object-js-components-web';
 import React, {useMemo} from 'react';
 
-import {ObjectFieldErrors} from '../ObjectFieldFormBase';
+import {ObjectFieldErrors} from './ObjectFieldFormBase';
 
 const defaultLanguageId = Liferay.ThemeDisplay.getDefaultLanguageId();
 const languages = Liferay.Language.available;
