+++ modules/apps/object/object-web/types/src/main/resources/META-INF/resources/js/components/ObjectField/AddObjectField.d.ts
@@ -12,7 +12,7 @@
  * details.
  */
 
-import React from 'react';
+/// <reference types="react" />
 
 export default function AddObjectField({
 	apiURL,
